% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData-package.R
\docType{package}
\name{scData-package}
\alias{scData-package}
\alias{_PACKAGE}
\title{scData: Sample Data for 'MultiK'}
\description{
Provides sample data for 'MultiK'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/siyao-liu/scData}
  \item Report bugs at \url{https://github.com/siyao-liu/scData/issues}
}

}
\author{
\strong{Maintainer}: Siyao Liu \email{siyao_liu@med.unc.edu}

Other contributors:
\itemize{
  \item Stuart R. Jefferys \email{srjefferys@gmail.com} (\href{https://orcid.org/0000-0001-6259-085X}{ORCID}) [contributor]
}

}
\keyword{internal}
