% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scData.R
\docType{data}
\name{scData}
\alias{scData}
\title{Sample scRNA data for use with MultiK}
\format{
\code{Seurat} object
See \href{https://satijalab.org/seurat/}{Seurat}

\strong{assays}
\itemize{
\item \strong{RNA} - Unique UMI counts, 2609 cells x 21247 features
}

\strong{cell data}
\itemize{
\item \strong{rownames} - (orig.ident)_(cell-barcode), e.g. "fibro_AAACGGGGTCTCGTTC".
\item \strong{orig.ident} - Sample origin, one of "fibro" (214 cells), "lum" (541),
or "basal" (1854).
\item \strong{nCount_RNA} - Number of RNA reads from cell (integer).
\item \strong{nFeature_RNA} - Number of unique UMIs = number of features in cell
(integer).
}

\strong{feature data}
\itemize{
\item \strong{rownames} - Genbank locus ID, e.g. "AL627309.1"
}
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM4041646}{GEO: GSM4041646}
}
\usage{
scData
}
\description{
A subset of data from the GEO data set
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM4041646}{GSM4041646}
processed with \href{https://satijalab.org/seurat/}{Seurat} and provided as an object of class
\href{https://cran.r-project.org/web/packages/SeuratObject/}{SeuratObject}.
The source data used to create this subset are the three Cell Ranger output
files from a scRNA analysis of a mixture of 3 human cell lines (MDA-MB-438,
MCF7, and Human dermal fibroblast (HF)).
}
\keyword{datasets}
