% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdata_add_data.R
\name{scdata_add_data}
\alias{scdata_add_data}
\title{Function to add single-cell data to the 'scdata' package for simple retrieval}
\usage{
scdata_add_data(
  package_path = package_path,
  file_name = file_name,
  sample_path = sample_path,
  sample_id = sample_id,
  file_type = "10x_h5",
  multi = FALSE,
  feature_type = "ENS",
  overwrite = TRUE
)
}
\arguments{
\item{package_path}{Path to 'scdata' package}

\item{file_name}{File name of input data}

\item{sample_path}{Path to input data (do not include file_name)}

\item{sample_id}{Name of sample for 'scdata' storage & retrieval (if 'file_name' is the 'sample_id', then add 'file_name')}

\item{file_type}{Input data file type used to select appropriate Read function (default is "10x_h5")}

\item{multi}{FALSE if data is NOT multi-modal (default); TRUE if data is multi-modal and input file contains multiple slots (only reads "Gene Expression" slot)}

\item{feature_type}{Define feature (row names) recovered from input data - if 10x GEX/RNA data, select "ENS" for Ensembl ID or "ID" for gene ID}

\item{overwrite}{If TRUE, overwrite data present in 'scdata' if Sample_ID already exists in 'scdata' library; If FALSE, there will be an Error flag if Sample_ID already exists in 'scdata' library}
}
\value{
Data added to the 'scdata' package
}
\description{
This function add single-cell data to the 'scdata' package for simple retrieval
}
\examples{
scdata_add_data(package_path=path/to/scdata, sample_id=sample_id, sample_path=path/to/sample_id, file_name=filtered_feature_bc_matrix.h5, file_type="10x_h5", multi=FALSE, feature_type="ENS")
}
