% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdata_add_metadata.R
\name{scdata_add_metadata}
\alias{scdata_add_metadata}
\title{Function to add sample MetaData to the 'scdata' package to review available data}
\usage{
scdata_add_metadata(
  package_path = package_path,
  annotation_file_name = annotation_file_name,
  annotation_path = annotation_path,
  print_sample_id = TRUE
)
}
\arguments{
\item{package_path}{Path to 'scdata' package}

\item{annotation_file_name}{Name of annotation file (likely "scdata_annotation.csv")}

\item{annotation_path}{Path to input annotation file (do not include file_name)}

\item{print_sample_id}{TRUE to print all sample IDs listed in annotation file; FALSE to NOT print sample IDs (Default = TRUE)}
}
\value{
MetaData added to the 'scdata' package
}
\description{
This function add single-cell data to the 'scdata' package for simple retrieval
}
\examples{
scdata_add_metadata(package_path=path/to/scdata, annotation_file_name=scdata_annotation, annotation_path=path/to/annotation, print_sample_id=TRUE)
}
